#1529 implemented requested changes moved multi search response classes to separate packages
