+++ cobigen/cobigen-core-api/src/main/java/com/devonfw/cobigen/api/util/to/nexus2/Nexus2SearchResponseArtifactHits.java
@@ -1,0 +1,22 @@
+package com.devonfw.cobigen.api.util.to.nexus2;
+
+import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ *
+ * Nexus search response artifacthits model
+ *
+ */
+@JsonIgnoreProperties(value = { "repositoryId" })
+class Nexus2SearchResponseArtifactHits {
+
+  /**
+   * artifactLinks
+   */
+  @JsonProperty("artifactLinks")
+  public List<Nexus2SearchResponeArtifactLinks> artifactLinks;
+
+}
